Fully norm proofed bzero and memset.

+++ ft_memset.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   ft_memset.c                                        :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: kgajadie <kgajadie@student.codam.nl>         +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2020/11/17 15:31:12 by kgajadie      #+#    #+#                 */
+/*   Updated: 2020/11/17 15:31:42 by kgajadie      ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned long	i;
+	size_t			i;
 	unsigned char	*new_ptr;
 
 	i = 0;
